requirements.txt
streamlit>=1.36.0
streamlit-authenticator>=0.4.1
pandas>=2.0.0
PyYAML>=6.0.0
